+++ app/components/developers/invisible_banner_component.rb
@@ -1,0 +1,21 @@
+module Developers
+  class InvisibleBannerComponent < ApplicationComponent
+    include ComponentWithIcon
+
+    attr_reader :user
+
+    def initialize(user)
+      @user = user
+    end
+
+    def render?
+      invisible?
+    end
+
+    private
+
+    def invisible?
+      user&.developer&.invisible?
+    end
+  end
+end